Moved Spawn logics to SpawnData. Made it so the TimeProgressor for the MobsEvent now waits for the spawn to be done before running the timer. (if a time was added that is) Made it so the TimeProgressor is used to end the event if no other progressors are added. Moved the Timeline code to the internal folder and the game specific code to the Timeline folder inside the specific folder.
