+++ Assets/_Core/Scripts/Internal/Timeline.meta
@@ -1,7 +1,7 @@
 fileFormatVersion: 2
-guid: 57b89c66d21efea4487c2087003dce0c
+guid: 001beb4ffedb28842984c6d555c8ff8a
 folderAsset: yes
-timeCreated: 1546448558
+timeCreated: 1546439794
 licenseType: Free
 DefaultImporter:
   externalObjects: {}
