+++ Assets/_Core/Scripts/Internal/Timeline/BaseTimelineEvent.cs.meta
