+++ Assets/_Core/Scripts/Internal/Timeline/TimelineEventSlot.cs.meta
