+++ Assets/_Core/Scripts/Specific/Timeline.meta
@@ -1,7 +1,7 @@
 fileFormatVersion: 2
-guid: 001beb4ffedb28842984c6d555c8ff8a
+guid: 57b89c66d21efea4487c2087003dce0c
 folderAsset: yes
-timeCreated: 1546439794
+timeCreated: 1546448558
 licenseType: Free
 DefaultImporter:
   externalObjects: {}
